Fix for missing access points in the list.
In some circumstances not all available access points in range were being listed, or only the connected AP was being found.

The problem seems to affect Win10 (maybe adapter specific) and is exacerbated by the passage of time since the connection was established and also by frequent polling of the APs (with APs slowly disappearing over time from the list).
To remedy this, its necessary to tell Windows to scan for APs prior to enumerating the APs (opening the Win10 WiFi connection list would actually resolve the problem temporarily).

This fix performs a scan during initialisation and optionally rescans automatically when retrieving the list of APs (if more than 60s have elapsed since the last scan).
Strictly speaking you should wait for a notification the scan completed before retrieving the APs - but this implementation works well enough in a polling situation.