Fix connection to access points containing Latin diacritic characters in their SSID name
SSIDs names now processed as UTF8 to preserve accented/diacritic/unicode characters.
The name is also now escaped before output to the XML profile.

The WLAN profile XML templates have been updated to UFT8 format and to include the hex ssid.
This was required for the connection to work with accented/diacritic/unicode characters.

# Conflicts:
#	SimpleWifi/ProfileFactory.cs